# Reduce overhead of creating / parsing / serializing trace IDs.